+++ app/src/main/java/campus/tech/kakao/map/model/datasource/AppDatabase.kt
@@ -1,9 +1,10 @@
-package campus.tech.kakao.map.model
+package campus.tech.kakao.map.model.datasource
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
+import campus.tech.kakao.map.model.SavedLocation
 
 @Database(entities = [SavedLocation::class], version = 1)
 abstract class AppDatabase: RoomDatabase() {
