+++ app/src/main/java/campus/tech/kakao/map/model/datasource/LocationApi.kt
@@ -4,29 +4,21 @@ import android.util.Log
 import campus.tech.kakao.map.model.Location
 import campus.tech.kakao.map.model.Location.Companion.toLocation
 import campus.tech.kakao.map.model.LocationDto
-import campus.tech.kakao.map.model.repository.KakaoAPI
+import campus.tech.kakao.map.model.SearchFromKeywordResponse
 import campus.tech.kakao.map.model.repository.RetrofitInstance
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 
-class LocationDataSource {
+class LocationApi {
     companion object{
         private const val RESULT_SIZE = 15
     }
 
     private val client = RetrofitInstance.getInstance().create(KakaoAPI::class.java)
 
-    suspend fun getLocations(keyword: String): List<Location> {
+    suspend fun getLocations(keyword: String): SearchFromKeywordResponse? {
         return withContext(Dispatchers.IO){
-            val response = client.searchFromKeyword(keyword, RESULT_SIZE)
-            val locationDtos: List<LocationDto> = response.body()?.documents ?: emptyList()
-            Log.d("jieun", "locationDtos: " + locationDtos)
-            toLocations(locationDtos)
+            client.searchFromKeyword(keyword, RESULT_SIZE).body()
         }
     }
-
-    private fun toLocations(locationDtos: List<LocationDto>) =
-        locationDtos.map {
-            it.toLocation()
-        }
 }
