+++ app/src/main/java/campus/tech/kakao/map/model/datasource/SavedLocationDao.kt
@@ -1,9 +1,10 @@
-package campus.tech.kakao.map.model
+package campus.tech.kakao.map.model.datasource
 
 import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.Query
+import campus.tech.kakao.map.model.SavedLocation
 
 @Dao
 interface SavedLocationDao {
