+++ app/src/test/java/campus/tech/kakao/map/SavedLocationApiTest.kt
@@ -3,7 +3,6 @@ package campus.tech.kakao.map
 import android.database.sqlite.SQLiteDatabase
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import campus.tech.kakao.map.model.LocationDbHelper
 import campus.tech.kakao.map.model.SavedLocation
 import org.junit.After
 import org.junit.Assert.assertEquals
@@ -12,7 +11,7 @@ import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class SavedLocationDataSourceTest {
+class SavedLocationApiTest {
     private lateinit var locationLocalDataSource: SavedLocationDataSource
     private lateinit var locationDbHelper: LocationDbHelper
     private lateinit var sqLiteDatabase: SQLiteDatabase
